+++ packages/supermassive/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @graphitation/supermassive
 
-<!-- This log was last generated on Mon, 09 Mar 2026 15:22:57 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 19 Mar 2026 18:12:21 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 3.15.3
+
+Thu, 19 Mar 2026 18:12:21 GMT
+
+### Patches
+
+- revert: errors on missing types/resolvers (dsamsonov@microsoft.com)
+
+## 3.15.2
+
+Tue, 17 Mar 2026 10:18:09 GMT
+
+### Patches
+
+- [HOTFIX] schema loader added to subscriptions (77059398+vejrj@users.noreply.github.com)
+
+## 3.15.1
+
+Mon, 16 Mar 2026 17:10:12 GMT
+
+### Patches
+
+- fix: allows for partial data in case resolver or type definition is missing (dsamsonov@microsoft.com)
 
 ## 3.15.0