+++ packages/webpack-loader/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @graphitation/webpack-loader
 
-<!-- This log was last generated on Mon, 09 Mar 2026 15:22:57 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 19 Mar 2026 18:12:21 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.1.23
+
+Thu, 19 Mar 2026 18:12:21 GMT
+
+### Patches
+
+- Bump @graphitation/supermassive to v3.15.3
+
+## 1.1.22
+
+Tue, 17 Mar 2026 10:18:09 GMT
+
+### Patches
+
+- Bump @graphitation/supermassive to v3.15.2
+
+## 1.1.21
+
+Mon, 16 Mar 2026 17:10:12 GMT
+
+### Patches
+
+- Bump @graphitation/supermassive to v3.15.1
 
 ## 1.1.20